my commit

+++ 0x00-hello_world/6-size.c
@@ -12,11 +12,11 @@ int main(void)
 	float f;
 	char c;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(I));
-	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(l));
-	printf("Size of a float: %lu byte(s)\n", (unsigned)sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", sizeof(c));
+	printf("Size of an int: %lu byte(s)\n", sizeof(i));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(I));
+	printf("Size of long long int: %lu byte(s)\n", sizeof(l));
+	printf("Size of a float: %lu byte(s)\n", sizeof(f));
 	
 	return (0);
 }
